requirements.txt
streamlit==1.32.0
requests==2.28.1
transformers==4.33.3
torch==2.1.2+cpu --extra-index-url https://download.pytorch.org/whl/cpu
python-dotenv==1.0.0
rake-nltk==1.0.6
nltk==3.8.1
gTTS==2.4.0